+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - `/users/update` endpoint added as `Iterable::Users#update`
 - `/users/{email}` endpoint added as `Iterable::Users#for_email`
 - `/users/byUserId/{userId}` endpoint added as `Iterable::Users#for_id`
+- `/users/getFields` endpoint added as `Iterable::Users#fields`
 
 ## v0.1.0 - 2017-10-04